feat: useSite to useSiteConfig
