+++ docs/valaxy.config.ts
@@ -1,0 +1,29 @@
+import { defineConfig } from 'valaxy'
+const safelist = [
+  'i-ri-home-line',
+]
+export default defineConfig({
+  vite: {
+    base: '/',
+  },
+  unocss: {
+    safelist,
+  },
+
+  markdown: {
+    blocks: {
+      tip: {
+        icon: 'i-carbon-thumbs-up',
+      },
+      warning: {
+        icon: 'i-carbon-warning-alt',
+      },
+      danger: {
+        icon: 'i-carbon-warning',
+      },
+      info: {
+        text: 'i-carbon-information',
+      },
+    },
+  },
+})
