+++ packages/valaxy-theme-yun/composables/config.ts
@@ -1,5 +1,5 @@
 import { computed } from 'vue'
-import { useSiteConfig } from 'valaxy'
+import { useSite } from 'valaxy'
 import type { YunTheme } from '../types'
 
 /**
@@ -7,6 +7,6 @@ import type { YunTheme } from '../types'
  * @returns
  */
 export function useThemeConfig<ThemeConfig = YunTheme.Config>() {
-  const config = useSiteConfig<ThemeConfig>()
+  const config = useSite<ThemeConfig>()
   return computed(() => config!.value.themeConfig)
 }
