+++ packages/valaxy-theme-yun/layouts/layout.vue
@@ -1,7 +1,7 @@
 <script lang="ts" setup>
-import { useSiteConfig } from 'valaxy'
+import { useSite } from 'valaxy'
 
-const config = useSiteConfig()
+const config = useSite()
 </script>
 
 <template>
