+++ packages/valaxy/client/composables/comments/twikoo.ts
@@ -2,7 +2,7 @@ import { isClient, useScriptTag } from '@vueuse/core'
 import { computed } from 'vue'
 import { useI18n } from 'vue-i18n'
 import { useRoute } from 'vue-router'
-import { useSiteConfig } from '../..'
+import { useSite } from '../..'
 
 declare global {
   interface Window {
@@ -24,7 +24,7 @@ declare global {
  * @param options
  */
 export function useTwikoo(options: {} = {}) {
-  const config = useSiteConfig()
+  const config = useSite()
   const cdnPrefix = computed(() => config.value.cdn.prefix)
 
   const route = useRoute()
